refactoring extract_audio script for better error logging and  management

+++ src/preprocess/extract_audio.py
@@ -11,21 +11,26 @@ input_filelist = np.loadtxt(args.input_file_list, delimiter=',', dtype=str)
 if os.path.exists(args.target_fold) == False:
     os.makedirs(args.target_fold)
 
-# first resample audio
-for i in range(input_filelist.shape[0]):
-    input_f = input_filelist[i]
-    ext_len = len(input_f.split('/')[-1].split('.')[-1])
-    video_id = input_f.split('/')[-1][:-ext_len-1]
-    output_f_1 = args.target_fold + '/' + video_id + '_intermediate.wav'
-    os.system('ffmpeg -i {:s} -vn -ar 16000 {:s}'.format(input_f, output_f_1)) # save an intermediate file
-
-# then extract the first channel
 for i in range(input_filelist.shape[0]):
     input_f = input_filelist[i]
     ext_len = len(input_f.split('/')[-1].split('.')[-1])
     video_id = input_f.split('/')[-1][:-ext_len-1]
     output_f_1 = args.target_fold + '/' + video_id + '_intermediate.wav'
     output_f_2 = args.target_fold + '/' + video_id + '.wav'
-    os.system('sox {:s} {:s} remix 1'.format(output_f_1, output_f_2))
-    # remove the intermediate file
-    os.remove(output_f_1)
+    if os.path.exists(output_f_2):
+        print(f"File {output_f_2} already exists, skipping")
+        continue
+    # first resample audio
+    if os.path.exists(output_f_1):
+        print(f"File {output_f_1} already exists, skipping")
+    else:
+        print(f"File {output_f_1} doesn't exists, creating")
+        os.system('ffmpeg -i {:s} -vn -ar 16000 {:s} > ffmpeg.log'.format(input_f, output_f_1)) # save an intermediate file
+
+    # then extract the first channel
+    if os.path.exists(output_f_1):
+        print(f"Creating {output_f_2}")
+        os.system('sox {:s} {:s} remix 1  > sox.log'.format(output_f_1, output_f_2))
+        os.remove(output_f_1)
+    else:
+        print("WARNING: Failure in creating", output_f_1)
